add selects contacts of list when it is created, add sweetalert to delete fn

+++ src/app/clientes/(privates)/correo/contactos/AccionesContacto.tsx
@@ -3,19 +3,34 @@ import useFetch from "@/app/hooks/useFetch";
 import { IContacto } from "@/interfaces/contactos.interfaces";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
+import Swal from "sweetalert2";
 
 const AccionesContacto = ({ id }: { id: number | string }) => {
    const router = useRouter();
    const { fetchCS, loading } = useFetch();
 
    const eliminarContacto = async () => {
-      const res = await fetchCS<IContacto>({
-         url: `/contactos/${id}`,
-         method: "DELETE",
-      });
-      if (res) {
-         router.refresh();
-      }
+
+      Swal.fire({
+         title: '¿Seguro quieres eliminar este contacto?',
+         showCancelButton: true,
+         confirmButtonText: 'Eliminar',
+         cancelButtonText: 'Cancelar',
+         confirmButtonColor: 'rgb(220,38,38) '
+       }).then(async (result) => {
+         /* Read more about isConfirmed, isDenied below */
+         if (result.isConfirmed) {
+            const res = await fetchCS<IContacto>({
+               url: `/contactos/${id}`,
+               method: "DELETE",
+            });
+            if (res) {
+               router.refresh();
+            }
+         } 
+       })
+
+      
    };
    return (
       <>
